+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 ## [Unreleased]
 ### Added
 - Much more detailed typing annotations in `pynopegl`
+- `RenderDisplace` node for texture displacement
 
 ### Changed
 - CSV export in the HUD now always prints floats in C locale instead of quoted